+++ unsequel/__init__.py
@@ -1,3 +1,3 @@
 """UNSeQueL: query data in logical execution order."""
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"